srcs/backend/requirements.txt
Django==5.1.4
psycopg2-binary==2.9.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
social-auth-app-django==5.4.2
Pillow==11.0.0
python-dotenv==1.0.1
django-cors-headers==4.6.0
django-redis==5.4.0